+++ openbrain/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "1.4.15"
+__version__ = "1.17.0"
 
 # import os
 # from pathlib import Path