+++ openbrain/orm/model_agent_config.py
@@ -24,6 +24,12 @@ class DefaultSettings(Enum):
                        "leadmo_stop_conversation",
                        "connect_with_agent",
                        "send_lead_to_crm",  # TODO get rid of this, it's only here to support a legacy name for this tool
+                       "get_current_time",
+                       "leadmo_get_simple_calendar_appointment_slots",
+                       "leadmo_create_appointment",
+                       "leadmo_create_contact",
+                       "lls_scrub_phone_number",
+                       "leadmo_get_contact_info_from_context",
                        "tester"
                        ]