+++ openbrain/tools/toolbox.py
@@ -13,6 +13,7 @@ from openbrain.util import get_logger
 from openbrain.tools.tool_leadmo_update_contact import OBToolLeadmoUpdateContact
 from openbrain.tools.tool_leadmo_stop_conversation import OBToolLeadmoStopConversation
 from openbrain.tools.tool_leadmo_create_contact import OBToolLeadmoCreateContact
+from openbrain.tools.tool_get_current_time import OBToolGetCurrentTime
 
 logger = get_logger()
 
@@ -25,6 +26,11 @@ class Toolbox:  # invoker
         "leadmo_stop_conversation": "OBToolLeadmoStopConversation",
         "leadmo_create_contact": "OBToolLeadmoCreateContact",
         "tester": "OBToolTester",
+        "get_current_time": "OBToolGetCurrentTime",
+        "leadmo_get_simple_calendar_appointment_slots": "OBToolLeadmoGetSimpleCalendarAppointmentSlots",
+        "leadmo_create_appointment": "OBToolLeadmoCreateAppointment",
+        "lls_scrub_phone_number": "OBToolLLSScrubPhoneNumberTool",
+        "leadmo_get_contact_info_from_context": "OBToolLeadmoGetContactInfoFromContext",
         "send_lead_to_crm": "OBToolLeadmoCreateContact", # TODO: get rid of this, it's only here to support a legacy name for this tool
 
     }
@@ -64,7 +70,7 @@ class Toolbox:  # invoker
                 self.register_obtool(obtool)
         else:
             # add the "do nothing" tool
-            obtool = self.available_tools["OBToolDoNothing"]
+            obtool = self.available_tools["OBToolGetCurrentTime"]
             self.callback_handler.register_ob_tool(obtool)
             self.register_obtool(obtool)